Fully automated update with version tags (#5)
* Update notify.py

Removed the message asking user to reload, added the placeholders/arguments for freqtrade version, nfi version and nfi update comment.

* Update update.sh

Added the retrieval of freqtrade version and nfi version before running the restart shell script

* New shell scripts 

restart.sh kills any open screens (caution it will kill them all if you're running multiple - this can be adjusted to just kill the screen with name 'freqtrade_nfi_bot') and start.sh creates the virtual environment and screen before starting the bot. Start is separated so that it can be run on crontab @restart

* Update README.md

Updated the instructions for the new files and automated process.

* Update README.md

Changed `chmod` to one line

* Update notify.py

Changed the message to include HTML tags e.g. <bold> for the headings and <code> for the content

* Update README.md

Added background and update instructions